refactor: remove circular imports
